// src/app/generated-code/model/booking.ts
/**
 * car_rental_backend API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0-SNAPSHOT
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { User } from './user';
import { Car } from './car';
import { Rating } from './rating';
import { BookingStatus } from './booking-status';


export interface Booking {
    id?: number;
    name: string;
    lastName: string;
    email: string;
    phone: string;
    startDate: string;
    endDate: string;
    timeStamp: string;
    bookingStatus: BookingStatus;
    total: number;
    rating?: Rating;
    car: Car;
    user: User;
}
export namespace Booking {
}