// src/app/generated-code/model/booking-status.ts
/**
 * car_rental_backend API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0-SNAPSHOT
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export type BookingStatus = 'PENDING' | 'ACTIVE' | 'COMPLETED' | 'CANCELED';

export const BookingStatus = {
    Pending: 'PENDING' as BookingStatus,
    Active: 'ACTIVE' as BookingStatus,
    Completed: 'COMPLETED' as BookingStatus,
    Canceled: 'CANCELED' as BookingStatus
};